refactor: reorganize wallet package directory structure
Implements a more organized directory structure for the wallet package:
- core/ - Core wallet interfaces and implementation
- chain/ - Chain-specific implementations
- utils/ - Utility and helper functions
- transaction/ - Transaction-related functionality
- plugins/ - Plugin implementations

Maintains the same public API through index.ts while improving code organization.
