+++ typescript/packages/wallets/src/transaction/gas-estimator.ts
@@ -1,7 +1,7 @@
 import { Client, Account, Contract, ABI, Address } from "@radiustechsystems/sdk";
-import { RadiusTransaction, TransactionSimulationResult } from "./types";
-import { GasEstimationError } from "./errors";
-import { WalletCache } from "./cache";
+import { RadiusTransaction, TransactionSimulationResult } from "../core/types";
+import { GasEstimationError } from "../utils/errors";
+import { WalletCache } from "../utils/cache";
 
 /**
  * Gas estimation service for EVM transactions
