+++ typescript/packages/wallets/src/utils/errors.ts
